change image_recognition to image-recognition to be in sync with all benchmarks
