+++ image-recognition/Dev/expectation_calc.py
