+++ image-recognition/Dev/image_recog.py
@@ -3,7 +3,6 @@
 
 import pandas as pd
 import numpy as np
-import sys
 from sklearn.datasets import fetch_openml
 from sklearn.model_selection import train_test_split
 from sklearn.decomposition import PCA
@@ -14,31 +13,14 @@ from qiskit import QuantumCircuit, Aer, transpile,execute
 from qiskit.circuit.library import RealAmplitudes
 from sklearn.metrics import accuracy_score,log_loss
 from qiskit.circuit import ParameterVector
-from qiskit.opflow.primitive_ops import PauliSumOp
-from qiskit.quantum_info import SparsePauliOp
-
-import expectation_calc_qcnn
+import expectation_calc
 
 global expectation_calc_method
 
 # change the below variable to True if you want to use expectation_calc.py file for expectation value calculation
 expectation_calc_method = True
 
-from qiskit.primitives import Estimator
-
-estimator = Estimator()
-
-def exp_cal(state):
-    
-    op = SparsePauliOp("IIIIZIII")
-
-    expectation_value = estimator.run(state, op).result().values
-
-# for shot-based simulation:
-    expectation_value = estimator.run(state, op, shots=100).result().values
-
-    return expectation_value
-
+debug = False
 # Dev Note :- Each image has Intensity from 0 to 255
 
 ''' All steps are explained below
@@ -69,8 +51,7 @@ def exp_cal(state):
     # Pending
     #   To Find common expectatation value code & other additional functionality for 8 qubits
     
-
-global num_qubits
+import expectation_calc
 
 # Number of qubits for the quantum circuit
 num_qubits = 8
@@ -130,6 +111,16 @@ pca = PCA(n_components = num_qubits)
 x_train_pca = pca.fit_transform(x_train)
 x_test_pca =  pca.fit_transform(x_test)
 
+# To visualize if prinicipal components are enough and to decide the number of principal components to keep 
+pca_check = False 
+if pca_check == True:
+    cumulative_variance_ratio = np.cumsum(pca.explained_variance_ratio_)
+
+    plt.plot(range(1, len(cumulative_variance_ratio) + 1), cumulative_variance_ratio)
+    plt.xlabel('Number of Principal Components')
+    plt.ylabel('Cumulative Explained Variance Ratio')
+    plt.show()
+
 
 # Create an instance of MinMaxScaler
 scaler = MinMaxScaler(feature_range=(0, 2 * np.pi))
@@ -139,112 +130,98 @@ scaler = MinMaxScaler(feature_range=(0, 2 * np.pi))
 x_scaled_test =  scaler.fit_transform(x_test_pca)
 x_scaled_train = scaler.fit_transform(x_train_pca)
 
+# ------------Dev Note Currently we are not using batching-----------------
+
+# Number of samples per batch
+samples_per_batch = 10
+
+# Calculate the total number of batches needed
+num_batches = np.ceil(len(x_scaled_train) / samples_per_batch)
+
+# Batch the x_scaled_data array
+x_scaled_test_batches = np.array_split(x_scaled_test, num_batches)
+
+# print(x_scaled_test_batches[0].shape)
+
+# for batch in x_scaled_test_batches[:3]:
+#     print("batch",batch)
+
+#----------- Dev Note Currently we are not using batching------------------
+
+# Data frame here to just to visualize the data and Not needed once model is defined
+pca_vis = False
+if pca_vis == True:
+    x_vis = pd.DataFrame(x_scaled_train)
+    stats = x_vis.describe()
+    print(stats)
 
 
-
-# Define the convolutional circuits
-def conv_circ_1(thetas, first, second):
-    # Your implementation for conv_circ_1 function here
-    # print(thetas)
-    conv_circ = QuantumCircuit(8)
-    conv_circ.rx(thetas[0], first)
-    conv_circ.rx(thetas[1], second)
-    conv_circ.rz(thetas[2], first)
-    conv_circ.rz(thetas[3], second)
-    conv_circ.crx(thetas[4], second, first)  
-    conv_circ.crx(thetas[5], first, second)
-    conv_circ.rx(thetas[6], first)
-    conv_circ.rx(thetas[7], second)
-    conv_circ.rz(thetas[8], first)
-    conv_circ.rz(thetas[9], second)
-    # print(conv_circ)
-    return conv_circ
-
-# Define the pooling circuits
-def pool_circ_1(thetas, first, second):
-    # Your implementation for pool_circ_1 function here
-    pool_circ = QuantumCircuit(8)
-    pool_circ.crz(thetas[0], first, second)
-    # pool_circ.x(second)
-    pool_circ.crx(thetas[1], first, second)
-    return pool_circ
-
-def pool_circ_2(first, second):
-    pool_circ = QuantumCircuit(8)
-    pool_circ.crz(first, second)
-    return pool_circ
-
-# Quantum Circuits for Convolutional layers
-def conv_layer_1(qc, thetas):
-    qc = qc.compose(conv_circ_1(thetas, 0, 1))
-    qc = qc.compose(conv_circ_1(thetas, 2, 3))
-    qc = qc.compose(conv_circ_1(thetas, 4, 5))
-    qc = qc.compose(conv_circ_1(thetas, 6, 7))
+# Variational circuit  used in below model which has parameters optimized during training
+# Dev Note will be replaced with QCNN 
+def var_circ(num_qubits, reps):
+    
+    # reps is Number of times ry and cx gates are repeated
+    qc = QuantumCircuit(num_qubits)
+    parameter_vector = ParameterVector("t", length=num_qubits*reps*2)
+    # print("parameter_vector",parameter_vector)
+    counter = 0
+    for rep in range(reps):
+        for i in range(num_qubits):
+            theta = parameter_vector[counter]
+            qc.ry(theta, i)
+            counter += 1
+        
+        for i in range(num_qubits):
+            theta = parameter_vector[counter]
+            qc.rx(theta, i)
+            counter += 1
+    
+        for j in range(0, num_qubits - 1, 1):
+            if rep<reps-1:
+                qc.cx(j, j + 1)
+    # print("counter",counter)
     return qc
 
-def conv_layer_2(qc, thetas):
-    qc = qc.compose(conv_circ_1(thetas, 0, 2))
-    qc = qc.compose(conv_circ_1(thetas, 4, 6))
-    return qc
 
-def conv_layer_3(qc, thetas):
-    qc = qc.compose(conv_circ_1(thetas, 0, 4))
-    return qc
+# var_circ for reference
+#      ┌──────────┐ ░            ░ ┌──────────┐ ░            
+# q_0: ┤ RY(θ[0]) ├─░───■────────░─┤ RY(θ[3]) ├─░───■────────
+#      ├──────────┤ ░ ┌─┴─┐      ░ ├──────────┤ ░ ┌─┴─┐      
+# q_1: ┤ RY(θ[1]) ├─░─┤ X ├──■───░─┤ RY(θ[4]) ├─░─┤ X ├──■───
+#      ├──────────┤ ░ └───┘┌─┴─┐ ░ ├──────────┤ ░ └───┘┌─┴─┐ 
+# q_2: ┤ RY(θ[2]) ├─░──────┤ X ├─░─┤ RY(θ[5]) ├─░──────┤ X ├─
+#      └──────────┘ ░      └───┘ ░ └──────────┘ ░      └───┘ 
 
-# Quantum Circuits for Pooling layers
-def pooling_layer_1(qc, thetas):
-    qc = qc.compose(pool_circ_1(thetas, 1, 0))
-    qc = qc.compose(pool_circ_1(thetas, 3, 2))
-    qc = qc.compose(pool_circ_1(thetas, 5, 4))
-    qc = qc.compose(pool_circ_1(thetas, 7, 6))
-    return qc
 
-def pooling_layer_2(qc, thetas):
-    qc = qc.compose(pool_circ_1(thetas, 0, 2))
-    qc = qc.compose(pool_circ_1(thetas, 4, 6))
-    return qc
-
-def pooling_layer_3(qc, thetas):
-    qc = qc.compose(pool_circ_1(thetas, 0, 4))
-    return qc
-
-debug = False
-# Variational circuit used in below model which has parameters optimized during training
-def qcnn_circ(num_qubits, thetas, layer_size = 10):
-    qc = QuantumCircuit(num_qubits,1)  # just to measure the 5 th qubit (4 indexed)\
-    if debug == True:
-        print(thetas)
-    theta1 = thetas[0:layer_size]
-    theta2 = thetas[layer_size: 2 * layer_size]
-    theta3 = thetas[2 * layer_size: 3 * layer_size]
-    theta4 = thetas[3 * layer_size: 3 * layer_size + 2]
-    theta5 = thetas[3 * layer_size + 2: 3 * layer_size + 4]
-    theta6 = thetas[3 * layer_size + 4: 3 * layer_size + 6]
-
-    # Pooling Ansatz1 is used by default
-    qc = conv_layer_1(qc, theta1)
-    qc = pooling_layer_1(qc, theta4)
-    qc = conv_layer_2(qc, theta2)
-    qc = pooling_layer_2(qc, theta5)
-    qc = conv_layer_3(qc, theta3)
-    qc = pooling_layer_3(qc, theta6)
-
-    return qc
 
 
 # model to be used for training which has input data encoded and variational circuit is appended to it
 def qcnn_model(theta, x, num_qubits, reps):
+    if expectation_calc_method == True:
+        qc = QuantumCircuit(num_qubits)
+    else:
+        qc = QuantumCircuit(num_qubits, num_qubits//2)
     
-    qc = QuantumCircuit(num_qubits)
-      
-    # Encode the pixel data into the quantum circuit here  x is the input data which is list of 14 values   
+    # Encode the pixel data into the quantum circuit here  x is the input data which is list of 14 values
+    
     # feature mapping 
     for j in range(num_qubits):
         qc.rx(x[j], j )
     # Append the variational circuit ( Ansatz ) to the quantum circuit
-    thetas = theta.tolist()
-    qcnn_circ_temp = qcnn_circ(num_qubits, thetas)
-    qc.compose(qcnn_circ_temp, inplace=True)
+    qcnn_circ = var_circ(num_qubits, reps)
+    qcnn_circ.assign_parameters(theta, inplace=True)
+    qc.compose(qcnn_circ, inplace=True)
+    # qc.measure_all()  # Measure all qubits will be changed to measure only 7 qubits if needed
+    # Measure only the first 7 qubits
+    # Add a classical register with 7 bits to store the results of the measurements
+    if expectation_calc_method == True:
+        # qc.measure_all()
+        pass
+    else:
+        for i in range(num_qubits//2):
+            # print("i",i, "num_qubits//2",num_qubits//2)
+            qc.measure(i, i)
+        
     return qc
 
 
@@ -255,7 +232,43 @@ backend = Aer.get_backend('qasm_simulator')
 global threshold 
 
 threshold = 0.5
+
+# function to calculate the expectation value ( pending :- Analysis in progress to improve or replace below function )
+def expectation_values(result):
+    expectation = 0
+    probabilities = {}
+    for outcome, count in result.items():
+        bitstring = outcome#[::-1]  # Reverse the bitstring
+        decimal_value = int(bitstring, 2)  # Convert the bitstring to decimal
+        probability = count / num_shots  # Compute the probability of each bitstring
+        expectation += decimal_value * probability / 2 ** (num_qubits/2)  # Compute the expectation value
+    if expectation > threshold:
+        prediction_label = 1
+    else:
+        prediction_label = 0
+    #print("expectation",expectation)
+    #return prediction_label
+
+prediction_label = []
+
+
+def extract_label(result):
+    bitstring = result['counts'].most_common(1)[0][0]
+    prediction_label = int(bitstring[::-1], 2)
+    if prediction_label == 0:
+        return 0
+    else:
+        return 1
     
+# function to calculate the loss function will be update after finding the suitable loss function
+ # Tried applying log loss ( cross entropy) but it is not minimizing the loss
+def square_loss(labels, predictions):
+    loss = 0
+    for l, p in zip(labels, predictions):
+        loss = loss + (l - p) ** 2
+    loss = loss / len(labels)
+    return loss
+
 
 # for training loss plot 
 
@@ -266,32 +279,30 @@ global num_iter_plot
 
 # function to calculate the loss function
 def loss_function(theta, is_draw_circ=False, is_print=True):
-
-    print("start----------------------------------------------------------------------------------")
-    print("theta",theta)
-    print("end------------------------------------------------------------------------------------")
-    
-    
+    # predicted_label = []
     prediction_label = []
     i_draw = 0
     for data_point, label in zip(x_scaled_train, y_train):
         # Create the quantum circuit for the data point
         qc = qcnn_model(theta, data_point, num_qubits, reps)
-        
+
         if i_draw==0 and is_draw_circ:
             print(qc)
             i_draw += 1
 
         # Simulate the quantum circuit and get the result
         if expectation_calc_method == True:
-            # val = expectation_calc_qcnn.calculate_expectation(qc,shots=num_shots,num_qubits=num_qubits) 
-            val = exp_cal(qc) 
-            # print(val)  
+            val = expectation_calc.calculate_expectation(qc,shots=num_shots,num_qubits=num_qubits)    
             val=(val+1)*0.5
-            # val = float(val[0])
-        prediction_label.append(float(val[0]))
-    
-    print(prediction_label)
+        else: 
+            job = backend.run(qc, shots=num_shots)
+            result = job.result().get_counts(qc)
+            val = expectation_values(result)
+        # predicted_label = extract_label(result)
+        #print("predicted_label",predicted_label)
+        prediction_label.append(val)
+    if debug == True:
+        print(prediction_label)
     loss = log_loss(y_train, prediction_label)
 
     if is_print:
@@ -301,7 +312,8 @@ def loss_function(theta, is_draw_circ=False, is_print=True):
                 predictions.append(1)
             else:
                 predictions.append(0)
-        # print(predictions)
+        if debug == True:
+            print(predictions)
         accuracy = accuracy_score(y_train, predictions)
         print("Accuracy:", accuracy, "loss:", loss)
         train_loss_history.append(loss)
@@ -309,30 +321,29 @@ def loss_function(theta, is_draw_circ=False, is_print=True):
     
     return loss
 
-# print(len(train_loss_history))
+
 
 def callback(theta):
     pass
-        # loss_function(theta, is_draw_circ=False, is_print=True)
 
 # Initialize  epochs
 num_epochs = 1
 reps = 3
+
 
 # Number of shots to run the program (experiment)
 num_shots = 1000
 
 # Initialize the weights for the QNN model
 np.random.seed(0)
-# weights = np.random.rand(num_qubits * reps *2 )
-weights = np.random.rand(36)
+weights = np.random.rand(num_qubits * reps *2 )
 #weights = np.zeros(num_qubits * reps)
 print(len(weights))
 
 # Will increase the number of epochs once the code is fine tuned to get convergance 
 for epoch in range(num_epochs):
     # Minimize the loss using SPSA optimizer
-    theta = minimize(loss_function, x0 = weights, method="COBYLA", tol=0.001, callback=callback, options={'maxiter': 100, 'rhobeg' : 0.5, 'disp': False} )
+    theta = minimize(loss_function, x0 = weights, method="COBYLA", tol=0.001, callback=callback, options={'maxiter': 150, 'disp': False} )
     #theta=SPSA(maxiter=100).minimize(loss_function, x0=weights)
 
     loss = theta.fun
@@ -340,6 +351,12 @@ for epoch in range(num_epochs):
 
 print("theta function", theta)
     # print(type(theta.x))
+# To find threshold
+# print("max_values",max_values)
+# max = max(max_values)
+# average_value = sum(max_values) / len(max_values)
+# print("average",average_value)
+# last step is to test the model on test data
 
 # Below code is to test the model on test data after training and getting theta values
 # Perform inference with the trained QNN
@@ -354,13 +371,17 @@ for data_point in x_scaled_test:
     qc = qcnn_model(theta.x, data_point, num_qubits, reps)
         # Simulate the quantum circuit and get the result
     if expectation_calc_method == True:
-        # val = expectation_calc_qcnn.calculate_expectation(qc,shots=num_shots,num_qubits=num_qubits)   
-        val = exp_cal(qc)   
-        val=(val+1)*0.5
+        val = expectation_calc.calculate_expectation(qc,shots=num_shots,num_qubits=num_qubits)  
+        val=(val+1)*0.5 
         if val > 0.5:
             predicted_label = 1
         else:
             predicted_label = 0
+    else: 
+        job = backend.run(qc, shots=num_shots)
+        result = job.result().get_counts(qc)
+        predicted_label = expectation_values(result)
+    # predicted_label = extract_label(result)
     print("predicted_label",predicted_label)
     predictions.append(predicted_label)
