+++ image-recognition/Dev/image_recog_QCNN 4 qubit Version.py
