+++ image-recognition/Dev/image_recog_QCNN.py
