+++ image-recognition/_common/instances/4_qubit.json
@@ -1,0 +1,1 @@
+[-0.07397172, 1.35735454, -0.3169999, 2.74608058, 2.57309511, 3.09098808, 1.51373312, -0.01892117, 0.08219012, 0.22553648, 0.48439645, 0.97254863, 0.95120939, 0.84353814, 0.36972275, 0.39970566, 0.22521085, -0.19281133, -0.36072423, 1.09389512, 0.8480923, 0.52705624, -0.21508983, 0.8841463]
