+++ image-recognition/qiskit/image_recognition_benchmark.py
